Dispatch EVENT_JOB_ERROR when job fails
The previous implementation of pool.py wraps the call to
f.result() in a try-except block - the problem is, previously
we would catch any exception raised by a job in ``run_job``,
so this exception would never propogate to this try/except block.
( I think)

In any case, the _run_job_error function was not dispatching
the EVENT_JOB_ERROR event, instead it just logged the exception.

We add a new function, ``handle_job_event()`` which determines
what action to take based on the event type returned by run_job().
This gives us a hook for other event types in the future